+++ packagedb/api.py
@@ -783,7 +783,7 @@ class CollectViewSet(viewsets.ViewSet):
       like with several Debian packages.
 
     **Examples::**
-    
+
         /api/collect/?purl=pkg:npm/foo@0.0.7&addon_pipelines=collect_symbols_ctags
 
         /api/collect/?purl=pkg:generic/busybox@1.36.1&addon_pipelines=collect_symbols_ctags&addon_pipelines=inspect_elf_binaries
@@ -818,6 +818,10 @@ class CollectViewSet(viewsets.ViewSet):
         purl = validated_data.get('purl')
 
         kwargs = dict()
+        # We want this request to have high priority since the user knows the
+        # exact package they want
+        kwargs['priority'] = 100
+
         if source_purl := validated_data.get('source_purl', None):
             kwargs["source_purl"] = source_purl
 
@@ -1005,9 +1009,11 @@ class CollectViewSet(viewsets.ViewSet):
                     # add to queue
                     extra_fields = dict()
                     if source_purl := package.get('source_purl'):
-                        extra_fields["source_uri"] = source_purl
+                        extra_fields['source_uri'] = source_purl
                     if addon_pipelines := package.get('addon_pipelines'):
-                        extra_fields["addon_pipelines"] = [pipe for pipe in addon_pipelines if is_supported_addon_pipeline(pipe)]
+                        extra_fields['addon_pipelines'] = [pipe for pipe in addon_pipelines if is_supported_addon_pipeline(pipe)]
+                    if priority := package.get('priority'):
+                        extra_fields['priority'] = priority
                     priority_resource_uri = PriorityResourceURI.objects.insert(purl, **extra_fields)
                     if priority_resource_uri:
                         queued_packages.append(purl)
@@ -1048,9 +1054,9 @@ class CollectViewSet(viewsets.ViewSet):
         If the package does not exist in the database this call does nothing.
         NOTE: this WILL NOT re-run scan and indexing in the background in contrast with the /collect
         and collect/index_packages endpoints.
-        
+
         **Request example**::
-        
+
             /api/collect/reindex_metadata/?purl=pkg:npm/foo@0.0.7
 
         """
@@ -1068,10 +1074,10 @@ class CollectViewSet(viewsets.ViewSet):
         packages = Package.objects.filter(**lookups)
         if packages.count() == 0:
             return Response(
-                {'status': f'Not recollecting: Package does not exist for {purl}'}, 
+                {'status': f'Not recollecting: Package does not exist for {purl}'},
                 status=status.HTTP_400_BAD_REQUEST,
             )
-        
+
         # Pass to only reindex_metadata downstream
         kwargs["reindex_metadata"] = True
         # here we have a package(s) matching our purl and we want to recollect metadata live
@@ -1234,6 +1240,8 @@ def get_resolved_packages(packages, supported_ecosystems):
             continue
 
         if parsed_purl.version:
+            # We prioritize Package requests that have explicit versions
+            package['priority'] = 100
             resolved_packages_by_purl[purl] = package
             continue